+++ app/actions/SeedServiceActions.js
@@ -1,4 +1,4 @@
-import { seeder } from '../middleware/grpc/seeder';
+import { seeder } from '../middleware/grpc/client';
 export const SEEDER_ATTEMPT = 'SEEDER_ATTEMPT';
 export const SEEDER_FAILED = 'SEEDER_FAILED';
 export const SEEDER_SUCCESS = 'SEEDER_SUCCESS';
