+++ app/actions/VersionActions.js
@@ -1,4 +1,4 @@
-import { getVersionService } from '../middleware/grpc/version';
+import { getVersionService } from '../middleware/grpc/client';
 import { loaderRequest } from './WalletLoaderActions';
 var messages = require('../middleware/walletrpc/api_pb');
 export const GETVERSIONSERVICE_ATTEMPT = 'GETVERSIONSERVICE_ATTEMPT';
